:hammer: Move test file to correct location
