+++ omisego-admin/src/liveTest/kotlin/co/omisego/omisego/live/BaseLiveTest.kt
@@ -1,4 +1,4 @@
-package co.omisego.omisego
+package co.omisego.omisego.live
 
 /*
  * OmiseGO
@@ -7,6 +7,7 @@ package co.omisego.omisego
  * Copyright © 2017-2018 OmiseGO. All rights reserved.
  */
 
+import co.omisego.omisego.OMGAPIAdmin
 import co.omisego.omisego.model.AdminConfiguration
 import co.omisego.omisego.network.ewallet.EWalletAdmin
 import co.omisego.omisego.utils.ResourceFileLoader
